Update methods for Gantt.java and Task.java
For Gantt: Added loadTasks and printTask to setup(). Changed load tasks to keep
track of index. This is so the task know where in the array list they
are and position themselves accordingly. Added border as argument in
currTask.display().

For Task: Added getters and setters. Added toString. Created display
method to display the names of the task on the left side of the gantt.