Change Gantt.java to accomadate mouse interaction
To increase or decrease task width the mouse has to be in the region
20px to the right and left of either end of the task. The left of the
task is called start and the right of the task is called end. To do
this in the event mousePressed() it goes through all tasks and checks
if the mouse X and Y are in any task's start or end region. If it is
call appropriate method from task to change width.

To know which task you have selected the current task that is selected
will be held in a variable called currTask. Once the mouse is released
the variable goes back to null so it is not affected again unless
selected again.