Add functionality to increase or decrease width
The way the tasks are drawn have been changed such that it uses
rectMode(CORNERS). To achieve this each vertices (corners) of the
task was recorded in vertices using the method getVertices().

Added a methods to detect which end of the task is being selected
(isStartArea() and isEndArea()). If the mouse is in any of the region
return true else return false.

To change the width of the tasks 2 methods were added, setStartVertices
and setEndVertices. When either is called the two vertices
corresponding to either start or end, the vertices are changed in set
amounts which is oneDayWidth. By changing the vertices the task is
redrawn using the new vertices which increases or decreases the task.

Task are not allowed to be smaller than one day. They cannot go past
the border of the lines either.